Move lib/client/BaseGraphqlSubscriber.js below lib/client/graphql/
