+++ lib/client/graphql/BaseGraphqlSubscriber.js
@@ -446,7 +446,7 @@ export default class BaseGraphqlSubscriber {
  * @typedef {{
  *   payload: furo.SubscriptionPayload<*> | null
  *   result: object | null
- *   abortedReason?: import('./BaseSubscriptionGraphqlCapsule').SUBSCRIBE_ABORTED_REASON
+ *   abortedReason?: import('../BaseSubscriptionGraphqlCapsule').SUBSCRIBE_ABORTED_REASON
  * }} CapsuleParams
  */
